[Windows] Refactor surface manager mocking
