+++ shell/platform/windows/testing/engine_modifier.h
@@ -29,14 +29,12 @@ class EngineModifier {
   // engine unless overwritten again.
   FlutterEngineProcTable& embedder_api() { return engine_->embedder_api_; }
 
-  // Explicitly sets the SurfaceManager being used by the FlutterWindowsEngine
-  // instance. This allows us to test fallback paths when a SurfaceManager fails
-  // to initialize for whatever reason.
+  // Override the surface manager used by the engine.
   //
   // Modifications are to the engine, and will last for the lifetime of the
   // engine unless overwritten again.
-  void SetSurfaceManager(AngleSurfaceManager* surface_manager) {
-    engine_->surface_manager_.reset(surface_manager);
+  void SetSurfaceManager(std::unique_ptr<AngleSurfaceManager> surface_manager) {
+    engine_->surface_manager_ = std::move(surface_manager);
   }
 
   /// Reset the start_time field that is used to align vsync events.
